// src/redux/levels-hidden-fields/levels-hidden-fields.types.ts
export interface ManualLevelsHiddenFields {
  subjectTypes: boolean;
  additionalPacks: boolean;
  customLevelPack: boolean;
  expectedSteps: boolean;
  expectedTime: boolean;
  levelNameEn: boolean;
  levelNameRu: boolean;
  levelCode: boolean;
  autoGeneratedLevelsCount: boolean;
  operations: boolean;
  stepsCountInterval: boolean;
  implicitTransformationsCount: boolean;
}

export interface AutoLevelsHiddenFields {
  startExpression: boolean;
  goalType: boolean;
  goalExpression: boolean;
  goalNaturalNumber: boolean;
  additionalPacks: boolean;
  customLevelPack: boolean;
  expectedSteps: boolean;
  expectedTime: boolean;
  levelNameEn: boolean;
  levelNameRu: boolean;
  levelCode: boolean;
}

export interface AllLevelsHiddenFields {
  manualLevelsHiddenFields: ManualLevelsHiddenFields;
  autoLevelsHiddenFields: AutoLevelsHiddenFields;
}

export enum LevelsHiddenFieldsActionTypes {
  TOGGLE_FIELD_VISIBILITY_FOR_ALL_MANUAL_LEVELS = "TOGGLE_FIELD_VISIBILITY_FOR_ALL_MANUAL_LEVELS",
  TOGGLE_FIELD_VISIBILITY_FOR_ALL_AUTO_LEVELS = "TOGGLE_FIELD_VISIBILITY_FOR_ALL_AUTO_LEVELS",
}